2기가 이상 파일 지원

+++ tdlnaService/inc/upnphttp.h
@@ -43,8 +43,8 @@ struct upnphttp {
 	int req_Timeout;
 	const char * req_SID; /* For UNSUBSCRIBE */
 	int req_SIDLen;
-	off_t req_RangeStart;
-	off_t req_RangeEnd;
+	unsigned long long req_RangeStart;
+	unsigned long long req_RangeEnd;
 	long int req_chunklen;
 	uint32_t reqflags;
